+++ packages/core/src/client.ts
@@ -3,27 +3,27 @@ import Event from './event'
 import Breadcrumb from './breadcrumb'
 import Session from './session'
 import includes from './lib/es-utils/includes'
-import filter from './lib/es-utils/filter'
-import reduce from './lib/es-utils/reduce'
 import assign from './lib/es-utils/assign'
 import runCallbacks from './lib/callback-runner'
 import metadataDelegate from './lib/metadata-delegate'
 import runSyncCallbacks from './lib/sync-callback-runner'
 import BREADCRUMB_TYPES from './lib/breadcrumb-types'
-import { add, clear, merge } from './lib/feature-flag-delegate'
+import featureFlagDelegate from './lib/feature-flag-delegate'
 import { BreadcrumbType, Config, Delivery, FeatureFlag, LoggerConfig, NotifiableError, Notifier, OnBreadcrumbCallback, OnErrorCallback, OnSessionCallback, Plugin, SessionDelegate, User } from './common'
+
+const HUB_PREFIX = '00000'
+const HUB_NOTIFY = 'https://notify.insighthub.smartbear.com'
+const HUB_SESSION = 'https://sessions.insighthub.smartbear.com'
 
 const noop = () => {}
 export default class Client<T extends Config = Config> {
   public readonly _notifier?: Notifier
   public readonly _config: T & Required<Config>
   private readonly _schema: any
-
-  public _delivery: Delivery
-
   private readonly _plugins: { [key: string]: any }
-
+  
   public _breadcrumbs: Breadcrumb[]
+  public _delivery: Delivery
   public _session: Session | null
   public _metadata: { [key: string]: any }
   public _featuresIndex: { [key: string]: number }
@@ -114,28 +114,28 @@ export default class Client<T extends Config = Config> {
     }
   }
 
-  addMetadata (section: string, keyOrObj: any, maybeVal?: any) {
+  addMetadata (section: string, keyOrObj: object | string, maybeVal?: any) {
     return metadataDelegate.add(this._metadata, section, keyOrObj, maybeVal)
   }
 
-  getMetadata (section: string, key?: any) {
+  getMetadata (section: string, key?: string) {
     return metadataDelegate.get(this._metadata, section, key)
   }
 
-  clearMetadata (section: string, key?: any) {
+  clearMetadata (section: string, key?: string) {
     return metadataDelegate.clear(this._metadata, section, key)
   }
 
   addFeatureFlag (name: string, variant: string | null = null) {
-    add(this._features, this._featuresIndex, name, variant)
+    featureFlagDelegate.add(this._features, this._featuresIndex, name, variant)
   }
 
   addFeatureFlags (featureFlags: FeatureFlag[]) {
-    merge(this._features, featureFlags, this._featuresIndex)
+    featureFlagDelegate.merge(this._features, featureFlags, this._featuresIndex)
   }
 
   clearFeatureFlag (name: string) {
-    clear(this._features, this._featuresIndex, name)
+    featureFlagDelegate.clear(this._features, this._featuresIndex, name)
   }
 
   clearFeatureFlags () {
@@ -156,7 +156,7 @@ export default class Client<T extends Config = Config> {
   }
 
   _configure (opts: T, internalPlugins: Plugin<T>[]) {
-    const schema = reduce(internalPlugins, (schema, plugin) => {
+    const schema = internalPlugins.reduce((schema, plugin) => {
       if (plugin && plugin.configSchema) return assign({}, schema, plugin.configSchema)
       return schema
     }, this._schema)
@@ -167,7 +167,7 @@ export default class Client<T extends Config = Config> {
     }
 
     // accumulate configuration and error messages
-    const { errors, config } = reduce(Object.keys(schema) as unknown as (keyof T)[], (accum, key: keyof typeof opts) => {
+    const { errors, config } = (Object.keys(schema) as (keyof T)[]).reduce((accum: {errors: Record<any, any>, config: Record<any, any>}, key: keyof typeof opts) => {
       const defaultValue = schema[key].defaultValue(opts[key])
 
       if (opts[key] !== undefined) {
@@ -194,11 +194,18 @@ export default class Client<T extends Config = Config> {
       if (!config.apiKey) throw new Error('No Bugsnag API Key set')
       // warn about an apikey that is not of the expected format
       if (!/^[0-9a-f]{32}$/i.test(config.apiKey)) errors.apiKey = 'should be a string of 32 hexadecimal characters'
+
+      if (opts.endpoints === undefined && config.apiKey.startsWith(HUB_PREFIX)) {
+        config.endpoints = {
+          notify: HUB_NOTIFY,
+          sessions: HUB_SESSION
+        }
+      }
     }
 
     // update and elevate some options
     this._metadata = assign({}, config.metadata)
-    merge(this._features, config.featureFlags, this._featuresIndex)
+    featureFlagDelegate.merge(this._features, config.featureFlags, this._featuresIndex)
     this._user = assign({}, config.user)
     this._context = config.context
     if (config.logger) this._logger = config.logger
@@ -213,7 +220,7 @@ export default class Client<T extends Config = Config> {
       this._logger.warn(generateConfigErrorMessage(errors, opts))
     }
 
-    return config
+    return config as T & Required<Config>
   }
 
   getUser () {
@@ -266,7 +273,7 @@ export default class Client<T extends Config = Config> {
   }
 
   removeOnError (fn: OnErrorCallback) {
-    this._cbs.e = filter(this._cbs.e, f => f !== fn)
+    this._cbs.e = this._cbs.e.filter( f => f !== fn)
   }
 
   _addOnSessionPayload (fn: OnSessionCallback) {
@@ -278,7 +285,7 @@ export default class Client<T extends Config = Config> {
   }
 
   removeOnSession (fn: OnSessionCallback) {
-    this._cbs.s = filter(this._cbs.s, f => f !== fn)
+    this._cbs.s = this._cbs.s.filter( f => f !== fn)
   }
 
   addOnBreadcrumb (fn: OnBreadcrumbCallback, front = false) {
@@ -286,7 +293,7 @@ export default class Client<T extends Config = Config> {
   }
 
   removeOnBreadcrumb (fn: OnBreadcrumbCallback) {
-    this._cbs.b = filter(this._cbs.b, f => f !== fn)
+    this._cbs.b = this._cbs.b.filter( f => f !== fn)
   }
 
   pauseSession () {
@@ -344,7 +351,7 @@ export default class Client<T extends Config = Config> {
     event._metadata = assign({}, event._metadata, this._metadata)
     event._user = assign({}, event._user, this._user)
     event.breadcrumbs = this._breadcrumbs.slice()
-    merge(event._features, this._features, event._featuresIndex)
+    featureFlagDelegate.merge(event._features, this._features, event._featuresIndex)
 
     // exit early if events should not be sent on the current releaseStage
     if (this._config.enabledReleaseStages !== null && !includes(this._config.enabledReleaseStages, this._config.releaseStage)) {
@@ -403,8 +410,7 @@ export default class Client<T extends Config = Config> {
 }
 
 const generateConfigErrorMessage = (errors: Record<string, Error>, rawInput: Config) => {
-  const er = new Error(
-  `Invalid configuration\n${(Object.keys(errors) as unknown as (keyof Config)[]).map(key => `  - ${key} ${errors[key]}, got ${stringify(rawInput[key])}`).join('\n\n')}`)
+  const er = new Error(`Invalid configuration\n${(Object.keys(errors) as unknown as (keyof Config)[]).map(key => `  - ${key} ${errors[key]}, got ${stringify(rawInput[key])}`).join('\n\n')}`)
   return er
 }
 
@@ -417,4 +423,3 @@ const stringify = (val: unknown) => {
     default: return String(val)
   }
 }
-