+++ packages/core/src/event.ts
@@ -1,11 +1,9 @@
 import { App, Device, FeatureFlag, Logger, Request, Stackframe, Thread, User, BugsnagError, NotifiableError } from "./common"
 
-import ErrorStackParser from './lib/error-stack-parser'
+import ErrorStackParser from 'error-stack-parser'
 // @ts-expect-error no types
 import StackGenerator from 'stack-generator'
 import hasStack from './lib/has-stack'
-import reduce from './lib/es-utils/reduce'
-import filter from './lib/es-utils/filter'
 import assign from './lib/es-utils/assign'
 import metadataDelegate from './lib/metadata-delegate'
 import featureFlagDelegate from './lib/feature-flag-delegate'
@@ -35,7 +33,7 @@ export default class Event {
   public threads: Thread[]
 
   public _metadata: { [key: string]: any }
-  public _features: FeatureFlag | null[]
+  public _features: (FeatureFlag | null)[]
   public _featuresIndex: { [key: string]: number }
 
   public _user: User
@@ -85,7 +83,7 @@ export default class Event {
     /* this.attemptImmediateDelivery, default: true */
   }
 
-  addMetadata (section: string, keyOrObj?: any, maybeVal?: any) {
+  addMetadata (section: string, keyOrObj?: object | string, maybeVal?: any) {
     return metadataDelegate.add(this._metadata, section, keyOrObj, maybeVal)
   }
 
@@ -165,7 +163,7 @@ static getStacktrace = function (error: Error, errorFramesToSkip: number, backtr
   if (hasStack(error)) return ErrorStackParser.parse(error).slice(errorFramesToSkip)
   // error wasn't provided or didn't have a stacktrace so try to walk the callstack
   try {
-    return filter(StackGenerator.backtrace(), (frame: StackTraceJsStyleStackframe) =>
+    return StackGenerator.backtrace().filter((frame: StackTraceJsStyleStackframe) =>
       (frame.functionName || '').indexOf('StackGenerator$$') === -1
     ).slice(1 + backtraceFramesToSkip)
   } catch (e) {
@@ -217,7 +215,7 @@ interface StackTraceJsStyleStackframe {
   args: string[],
   fileName: string,
   lineNumber: number,
-  columnNumber: number, 
+  columnNumber: number,
   isEval: boolean,
   isNative: boolean,
   source: string,
@@ -259,7 +257,7 @@ function createBugsnagError (errorClass: unknown, errorMessage: unknown, type: s
     errorClass: ensureString(errorClass),
     errorMessage: ensureString(errorMessage),
     type,
-    stacktrace: reduce(stacktrace, (accum, frame) => {
+    stacktrace: stacktrace.reduce((accum, frame) => {
       const f = formatStackframe(frame)
       // don't include a stackframe if none of its properties are defined
       try {