+++ packages/core/src/lib/es-utils/includes.js
@@ -1,4 +1,3 @@
-const reduce = require('./reduce')
 // Array#includes
 module.exports = (arr, x) =>
-  reduce(arr, (accum, item, i, arr) => accum === true || item === x, false)
+  arr.reduce((accum, item, i, arr) => accum === true || item === x, false)