+++ packages/plugin-inline-script-content/rollup.config.npm.mjs
@@ -2,9 +2,5 @@ import createRollupConfig from '../../.rollup/index.mjs'
 
 export default createRollupConfig({
     input: 'src/inline-script-content.ts',
-    external: [
-        '@bugsnag/core',
-        '@bugsnag/core/lib/es-utils/reduce',
-        '@bugsnag/core/lib/es-utils/filter'
-    ]
+    external: ['@bugsnag/core']
 })